Reorganizing code. Still going to be a significant mess.
I might have to change things around yet more. As things are,
I end up having to carry (and probably save) two copies for the
GetProject and GetArtifacts, where the uuid is actually the same.
A bit of a waste!

Further, I cannot use the ArtifactSha as a source of id; I'll need
some sort of File wrapper instead. Additionally, the package
distributed.repo.core is not visible where most of the config
class files are defined (d-metadata), which means I either have
to move the GetKeys elsewhere, or keep using the uuids as strings,
which however defeats the whole charade I've been doing with the
type safe access keys.