Refactoring ArtifactsOut
We are changing what is being returned by runBuild() from a
BuildArtifactsOut into an ArtifactsOut, which is similar but
does not contain a sha. That simplifies the code, and will
eventually allow us to encapsulate the sha calculation stuff
entirely inside the repo implementation code, which is where
it really belong: it should not be used elsewhere.

In principle, we could store in an ArtifactsOut directly a
list of Files, and lift the relativization code up to
LocalBuildRunner.runLocalBuild(). However, we do need to
be able to serialize ArtifactsOut, due to it being passed
back from sbt to the sbt build system, so we relativize
inside each build system. For now.