Enforcing encapsulation of GetKeys and Sections
By making GetKeys and Sections opaque, the client code in all the
rest of dbuild can be certain to be no longer impacted by changes
in the metadata structures inside the repository. Although that may
appear to be a bit strict, it allows for additional flexibility and
decoupling in future definition of the repository structure.