Better separation between core and user code
"core" is the core implementation of the Repository.
"user" is the interface that knows how to deal with metadata
and how to prepare it in a form that the Repository will store

The client code in the rest of dbuild can use the (opaque)
GetKeys and Sections, and the general calls in Repository
and ReadableRepository.

This concludes the refactoring that was useful to introduce
layering in the repository data access.

The next step is completing the low-level implementation of
the local/remote/cache repositories, and the related locking,
all of which is now going to be perfectly transparent to all
client code.